+++ lib/olels/objects/workbench_line.rb
@@ -17,31 +17,13 @@ module OLE_QA::Framework::OLELS
   class Workbench_Line < OLE_QA::Framework::Line_Object
     def set_elements
       super
-      element(:search_field)                {(line_id == 0) ?
-                                              id_str = 'SearchConditions_SearchText_id_add_control' :
-                                              id_str = "SearchConditions_SearchText_id_line#{line_id - 1}_control"
-                                              b.text_field(:id => id_str)}
+      element(:search_field)                {b.text_field(:id => "SearchConditions_SearchText_id_line#{line_id}_control")}
       # @note Options for this selector are 'AND', 'OR', and 'phrase'.
-      element(:search_scope_selector)       {(line_id == 0) ?
-                                              id_str = 'SearchConditions_Operator_id_add_control' :
-                                              id_str = "SearchConditions_Operator_id_line#{line_id - 1}_control"
-                                              b.select_list(:id => id_str)}
-      element(:field_selector)              {(line_id == 0) ?
-                                              id_str = 'SearchConditions_DocField_id_add_control' :
-                                              id_str = "SearchConditions_DocField_id_line#{line_id - 1}_control"
-                                              b.select_list(:id => id_str)}
-      element(:join_and)                    {(line_id == 0) ?
-                                              id_str = 'SearchConditions_SearchScope_id_add_control_0' :
-                                              id_str = "SearchConditions_SearchScope_id_line#{line_id - 1}_control_0"
-                                              b.radio(:id => id_str)}
-      element(:join_or)                     {(line_id == 0) ?
-                                              id_str = 'SearchConditions_SearchScope_id_add_control_1' :
-                                              id_str = "SearchConditions_SearchScope_id_line#{line_id - 1}_control_1"
-                                              b.radio(:id => id_str)}
-      element(:join_not)                    {(line_id == 0) ?
-                                              id_str = 'SearchConditions_SearchScope_id_add_control_2' :
-                                              id_str = "SearchConditions_SearchScope_id_line#{line_id - 1}_control_2"
-                                              b.radio(:id => id_str)}
+      element(:search_scope_selector)       {b.select_list(:id => "SearchConditions_Operator_id_line#{line_id}_control")}
+      element(:field_selector)              {b.select_list(:id => "SearchConditions_DocField_id_line#{line_id}_control")}
+      element(:join_and)                    {b.radio(:id => "SearchConditions_SearchScope_id_line#{line_id}_control_0")}
+      element(:join_or)                     {b.radio(:id => "SearchConditions_SearchScope_id_line#{line_id}_control_1")}
+      element(:join_not)                    {b.radio(:id => "SearchConditions_SearchScope_id_line#{line_id}_control_2")}
       element(:add_button)                  {b.button(:id => "addLineField-Add_line#{line_id}")}
       element(:delete_button)               {b.button(:id => "deleteLineField-Delete_line#{line_id}")}
     end